refactor: rename update consume media to record, convert to factory function that accepts the title field
